+++ gulpfile.js
@@ -3,7 +3,7 @@ const config = require('./gulp.config.js')();
 const fs = require('fs');
 const semver = require('semver');
 const argv = require('yargs').argv;
-const sass = require('gulp-sass');
+const sass = require('gulp-sass')(require('sass'));
 const del = require('del');
 const tildeImporter = require('node-sass-tilde-importer');