+++ src/lib/components/forms/form/form-components/text-area/text-area.interface.ts
@@ -4,4 +4,5 @@ export interface TextAreaInterface extends TerraFormComponentBaseInterface {
     hasFixedHeight: boolean;
     maxRows: number;
     maxLength: number;
+    minLength: number;
 }