+++ src/lib/translations/l10n-terra-components.ts
@@ -145,6 +145,11 @@ const de: {} = {
     },
     validators: {
         required: 'Dies ist ein Pflichtfeld.',
+        pattern: 'Eingabe entspricht nicht dem Schema: {{ pattern }}.',
+        patternDecimalCount: 'Anzahl der erforderlichen Nachkommastellen: {{ decimalCount }}.',
+        patternHex: 'Ungültiger Hex-Wert',
+        maxLength: 'Maximal {{maxLength}} Zeichen erlaubt.',
+        minLength: 'Mindestens {{minLength}} Zeichen benötigt.',
         min: 'Der Mindestwert ist unterschritten.',
         max: 'Der Maximalwert ist überschritten.'
     }
@@ -296,6 +301,11 @@ const en: {} = {
     },
     validators: {
         required: 'This is a mandatory field.',
+        pattern: 'Input does not match the pattern: {{ pattern }}.',
+        patternDecimalCount: 'Number of required decimal places: {{ decimalCount }}.',
+        patternHex: 'Invalid hex value',
+        maxLength: 'A maximum of {{maxLength}} characters is allowed.',
+        minLength: 'A minimum of {{minLength}} characters is required.',
         min: 'The minimum value has been exceeded.',
         max: 'The maximum value has been exceeded.'
     }